refactor(post&category)&fix(controller): 优化 post 和 category 模块; 新增 post_categories 中间表; 修复 controller 失败响应的格式问题

+++ configs/config.go
@@ -40,8 +40,8 @@ type LogConfig struct {
 	LogFilePath     string `mapstructure:"LOG_FILE_PATH"`
 	LogFileName     string `mapstructure:"LOG_FILE_NAME"`
 	LogTimestampFmt string `mapstructure:"LOG_TIMESTAMP_FMT"`
-	LogMaxAge       int    `mapstructure:"LOG_MAX_AGE"`
-	LogRotationTime int    `mapstructure:"LOG_ROTATION_TIME"`
+	LogMaxAge       int64  `mapstructure:"LOG_MAX_AGE"`
+	LogRotationTime int64  `mapstructure:"LOG_ROTATION_TIME"`
 	LogLevel        string `mapstructure:"LOG_LEVEL"`
 }
